requirements.txt
python-dateutil==2.7.3
python-whois==0.7.0
requests==2.11.1
